refactor: Member 모듈로 통합
과도하게 분리된 모듈 정리
