+++ server/src/main/java/com/emr/slgi/member/controller/MemberController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.emr.slgi.member.MemberService;
+import com.emr.slgi.member.service.MemberService;
 
 import lombok.RequiredArgsConstructor;
