+++ server/src/main/java/com/emr/slgi/member/controller/StaffController.java
@@ -1,4 +1,4 @@
-package com.emr.slgi.staff;
+package com.emr.slgi.member.controller;
 
 import java.util.Map;
 
@@ -7,6 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.emr.slgi.member.service.StaffService;
 
 import lombok.RequiredArgsConstructor;
