--- server/src/main/java/com/emr/slgi/nurse/NurseController.java
@@ -1,19 +1,0 @@
-package com.emr.slgi.nurse;
-
-import com.emr.slgi.reservation.dto.ReservationForm;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Map;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/nurse")
-public class NurseController {
-
-
-}
